moving fixtures and preparing test
